fix: change short description prompt to not allow empty values
The current implementation of the short description prompt do not output a validation message
when the user input is empty. As this is a required value as per the standard-readme specs,
we need to disallow empty values for the short description. This commits adds the length
validator to the prompt to disallow empty inputs.

fixes #48