docs: add a known issue section in readme with information of hang issue
There is an issue on windows where generators hang either during the running, or just after the
generator have run.

This is a problem with yeoman itself (or one of the dependencies), and is not
fixable in the cake-addin generator. As such creating a section in the readme informing of this
known issue, and what can be done when it happens is appropriate.

NOTE: thanks to @gep13 for finding this issue.