feat: remove the appveyor and travis generator from the default generator
This commit makes the changes necessary to the default (app) generator and the different template
files, to ensure that the appveyor and travis generators no longer are being runned when just
calling `yo cake-addin`.

BREAKING CHANGE: The ability to generate an appveyor or travis build script have been removed from
the default generator (the one used when just calling `yo cake-addin`). This change was made as we
do not want to force anyone to use one or more CI providers. To generate an appveyor or travis
script can still be done by calling one of the sub-generators (`yo cake-addin:appveyor` and `yo
cake-addin:travis`), but there will still no longer be any badges in the Readme.md for these
providers anymore.